requirements.txt
streamlit
google_generativeAi
python-dotenv
langchain